MAINT: rename packages
