+++ data-prepper-plugins/otel-trace-group-processor/src/main/java/com/amazon/dataprepper/plugins/processor/oteltracegroup/model/TraceGroup.java
@@ -1,9 +1,21 @@
-package com.amazon.dataprepper.plugins.prepper.oteltrace.model;
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package com.amazon.dataprepper.plugins.processor.oteltracegroup.model;
 
 import com.amazon.dataprepper.model.trace.Span;
 import com.amazon.dataprepper.model.trace.TraceGroupFields;
 
+import java.util.Objects;
+
 public class TraceGroup {
+    public static final String TRACE_GROUP_NAME_FIELD = "traceGroup";
+    public static final String TRACE_GROUP_END_TIME_FIELD = "traceGroupFields.endTime";
+    public static final String TRACE_GROUP_STATUS_CODE_FIELD = "traceGroupFields.statusCode";
+    public static final String TRACE_GROUP_DURATION_IN_NANOS_FIELD = "traceGroupFields.durationInNanos";
+
     private final String traceGroup;
 
     private final TraceGroupFields traceGroupFields;
@@ -16,6 +28,19 @@ public class TraceGroup {
         return traceGroupFields;
     }
 
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        TraceGroup that = (TraceGroup) o;
+        return Objects.equals(traceGroup, that.traceGroup) && Objects.equals(traceGroupFields, that.traceGroupFields);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(traceGroup, traceGroupFields);
+    }
+
     TraceGroup(final TraceGroupBuilder traceGroupBuilder) {
         traceGroup = traceGroupBuilder.traceGroup;
         traceGroupFields = traceGroupBuilder.traceGroupFields;
@@ -25,12 +50,12 @@ public class TraceGroup {
         private String traceGroup;
         private TraceGroupFields traceGroupFields;
 
-        public TraceGroupBuilder setTraceGroup(final String traceGroup) {
+        public TraceGroupBuilder setTraceGroup(String traceGroup) {
             this.traceGroup = traceGroup;
             return this;
         }
 
-        public TraceGroupBuilder setTraceGroupFields(final TraceGroupFields traceGroupFields) {
+        public TraceGroupBuilder setTraceGroupFields(TraceGroupFields traceGroupFields) {
             this.traceGroupFields = traceGroupFields;
             return this;
         }
