+++ data-prepper-plugins/otel-trace-group-processor/src/test/java/com/amazon/dataprepper/plugins/processor/oteltracegroup/OTelTraceGroupProcessorConfigTests.java
@@ -1,15 +1,9 @@
 /*
- *  SPDX-License-Identifier: Apache-2.0
- *  
- *  The OpenSearch Contributors require contributions made to
- *  this file be licensed under the Apache-2.0 license or a
- *  compatible open source license.
- *
- *  Modifications Copyright OpenSearch Contributors. See
- *  GitHub history for details.
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.amazon.dataprepper.plugins.prepper.oteltracegroup;
+package com.amazon.dataprepper.plugins.processor.oteltracegroup;
 
 import com.amazon.dataprepper.model.configuration.PluginSetting;
 import com.amazon.dataprepper.plugins.sink.opensearch.ConnectionConfiguration;
