--- data-prepper-plugins/otel-trace-raw-processor/src/main/java/com/amazon/dataprepper/plugins/prepper/oteltrace/OtelTraceRawProcessorConfig.java
@@ -1,19 +1,0 @@
-/*
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  The OpenSearch Contributors require contributions made to
- *  this file be licensed under the Apache-2.0 license or a
- *  compatible open source license.
- *
- *  Modifications Copyright OpenSearch Contributors. See
- *  GitHub history for details.
- */
-
-package com.amazon.dataprepper.plugins.prepper.oteltrace;
-
-public class OtelTraceRawProcessorConfig {
-    static final String TRACE_FLUSH_INTERVAL = "trace_flush_interval";
-    static final long DEFAULT_TG_FLUSH_INTERVAL_SEC = 180L;
-    static final long DEFAULT_TRACE_ID_TTL_SEC = 15L;
-    static final long MAX_TRACE_ID_CACHE_SIZE = 1000_000L;
-}
