+++ data-prepper-plugins/otel-trace-raw-processor/src/main/java/com/amazon/dataprepper/plugins/processor/oteltrace/model/TraceGroup.java
@@ -1,16 +1,14 @@
-package com.amazon.dataprepper.plugins.prepper.oteltracegroup.model;
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package com.amazon.dataprepper.plugins.processor.oteltrace.model;
 
 import com.amazon.dataprepper.model.trace.Span;
 import com.amazon.dataprepper.model.trace.TraceGroupFields;
 
-import java.util.Objects;
-
 public class TraceGroup {
-    public static final String TRACE_GROUP_NAME_FIELD = "traceGroup";
-    public static final String TRACE_GROUP_END_TIME_FIELD = "traceGroupFields.endTime";
-    public static final String TRACE_GROUP_STATUS_CODE_FIELD = "traceGroupFields.statusCode";
-    public static final String TRACE_GROUP_DURATION_IN_NANOS_FIELD = "traceGroupFields.durationInNanos";
-
     private final String traceGroup;
 
     private final TraceGroupFields traceGroupFields;
@@ -23,19 +21,6 @@ public class TraceGroup {
         return traceGroupFields;
     }
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        TraceGroup that = (TraceGroup) o;
-        return Objects.equals(traceGroup, that.traceGroup) && Objects.equals(traceGroupFields, that.traceGroupFields);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(traceGroup, traceGroupFields);
-    }
-
     TraceGroup(final TraceGroupBuilder traceGroupBuilder) {
         traceGroup = traceGroupBuilder.traceGroup;
         traceGroupFields = traceGroupBuilder.traceGroupFields;
@@ -45,12 +30,12 @@ public class TraceGroup {
         private String traceGroup;
         private TraceGroupFields traceGroupFields;
 
-        public TraceGroupBuilder setTraceGroup(String traceGroup) {
+        public TraceGroupBuilder setTraceGroup(final String traceGroup) {
             this.traceGroup = traceGroup;
             return this;
         }
 
-        public TraceGroupBuilder setTraceGroupFields(TraceGroupFields traceGroupFields) {
+        public TraceGroupBuilder setTraceGroupFields(final TraceGroupFields traceGroupFields) {
             this.traceGroupFields = traceGroupFields;
             return this;
         }
